refactor(core-webhooks): drop no longer needed pagination
