--- packages/core-json-rpc/src/server/modules/index.ts
@@ -1,5 +1,0 @@
-import * as blocks from "./blocks";
-import * as transactions from "./transactions";
-import * as wallets from "./wallets";
-
-export { blocks, transactions, wallets };
