+++ packages/core-webhooks/src/server/utils.ts
@@ -1,6 +1,7 @@
 import Boom from "boom";
+import { IWebhook } from "../interfaces";
 
-export function transformResource(model) {
+export function transformResource(model): IWebhook {
     return {
         id: model.id,
         event: model.event,
@@ -11,20 +12,6 @@ export function transformResource(model) {
     };
 }
 
-export function paginate(request) {
-    return {
-        offset: (request.query.page - 1) * request.query.limit,
-        limit: request.query.limit,
-    };
-}
-
 export function respondWithResource(data) {
     return data ? { data: transformResource(data) } : Boom.notFound();
-}
-
-export function toPagination(data) {
-    return {
-        results: data.rows.map(d => transformResource(d)),
-        totalCount: data.count,
-    };
 }
